Move pages classes into core items namespace
And improve general wording of our page classes
