+++ UnityPomodoro/Assets/AdrianMiasik/Scripts/Components/Base/SettingsOptionDropdown.cs
@@ -1,5 +1,5 @@
-﻿using AdrianMiasik.Components.Specific;
-using AdrianMiasik.Components.Specific.Pages;
+﻿using AdrianMiasik.Components.Core.Items.Pages;
+using AdrianMiasik.Components.Specific;
 using AdrianMiasik.Components.Specific.Settings;
 using AdrianMiasik.ScriptableObjects;
 using TMPro;
