+++ UnityPomodoro/Assets/AdrianMiasik/Scripts/Components/Base/SettingsOptionToggleSlider.cs
@@ -1,7 +1,7 @@
 using AdrianMiasik.Components.Core;
+using AdrianMiasik.Components.Core.Items.Pages;
 using AdrianMiasik.Components.Core.Settings;
 using AdrianMiasik.Components.Specific;
-using AdrianMiasik.Components.Specific.Pages;
 using AdrianMiasik.Components.Specific.Settings;
 using AdrianMiasik.ScriptableObjects;
 using TMPro;
