+++ UnityPomodoro/Assets/AdrianMiasik/Scripts/Components/Core/Items/Pages/TimerPage.cs
@@ -3,8 +3,11 @@ using AdrianMiasik.Components.Core.Containers;
 using AdrianMiasik.ScriptableObjects;
 using UnityEngine;
 
-namespace AdrianMiasik.Components.Specific.Pages
+namespace AdrianMiasik.Components.Core.Items.Pages
 {
+    /// <summary>
+    /// Our main timer <see cref="Page"/> that is used to show/hide the digit and timer content.
+    /// </summary>
     public class TimerPage: Page
     {
         [SerializeField] private DigitFormat m_format;
