+++ UnityPomodoro/Assets/AdrianMiasik/Scripts/Components/Specific/Settings/OptionMuteAudioOutOfFocus.cs
@@ -1,6 +1,6 @@
 ﻿using AdrianMiasik.Components.Base;
+using AdrianMiasik.Components.Core.Items.Pages;
 using AdrianMiasik.Components.Core.Settings;
-using AdrianMiasik.Components.Specific.Pages;
 
 namespace AdrianMiasik.Components.Specific.Settings
 {
