+++ UnityPomodoro/Assets/AdrianMiasik/Scripts/Components/Specific/Settings/OptionUnityAnalytics.cs
@@ -1,8 +1,8 @@
 using System.Collections.Generic;
 using AdrianMiasik.Components.Base;
 using AdrianMiasik.Components.Core;
+using AdrianMiasik.Components.Core.Items.Pages;
 using AdrianMiasik.Components.Core.Settings;
-using AdrianMiasik.Components.Specific.Pages;
 using Unity.Services.Analytics;
 using UnityEngine;
