+++ UnityPomodoro/Assets/AdrianMiasik/Scripts/PomodoroTimer.cs
@@ -7,7 +7,6 @@ using AdrianMiasik.Components.Core.Containers;
 using AdrianMiasik.Components.Core.Items;
 using AdrianMiasik.Components.Core.Settings;
 using AdrianMiasik.Components.Specific;
-using AdrianMiasik.Components.Specific.Pages;
 using AdrianMiasik.Interfaces;
 using AdrianMiasik.ScriptableObjects;
 using AdrianMiasik.UWP;
